#completed salary.

+++ VINASIC.Business.Interface/Model/ModelUser.cs
@@ -61,7 +61,7 @@ namespace VINASIC.Business.Interface.Model
         public int? Index { get; set; }
         public string Content { get; set; }
         [System.ComponentModel.DefaultValue(0)]
-        public float? Amount { get; set; }
+        public double? Amount { get; set; }
         public string Unit { get; set; }
     }
 }
